+++ docling_jobkit/datamodel/task.py
@@ -10,10 +10,11 @@ from docling.datamodel.document import ConversionResult
 
 from docling_jobkit.datamodel.convert import ConvertDocumentsOptions
 from docling_jobkit.datamodel.http_inputs import FileSource, HttpSource
+from docling_jobkit.datamodel.s3_coords import S3Coordinates
 from docling_jobkit.datamodel.task_meta import TaskProcessingMeta, TaskStatus
 from docling_jobkit.datamodel.task_targets import InBodyTarget, TaskTarget
 
-TaskSource = Union[HttpSource, FileSource, DocumentStream]
+TaskSource = Union[HttpSource, FileSource, DocumentStream, S3Coordinates]
 
 
 class Task(BaseModel):