+++ docling_jobkit/kfp_pipeline/docling_s3in_s3out.py
@@ -1,6 +1,6 @@
 # ruff: noqa: E402
 
-from typing import List, NamedTuple, cast
+from typing import List, NamedTuple
 
 from kfp import dsl
 from kfp.dsl import Dataset, Input, Output
@@ -9,13 +9,12 @@ from kfp.dsl import Dataset, Input, Output
 @dsl.component(
     packages_to_install=[
         "docling==2.28.0",
-        "git+https://github.com/docling-project/docling-jobkit@2c27c71b75da98f04fccc7abc7ddc3a9a3afb0cd",
+        "git+https://github.com/docling-project/docling-jobkit@main",
     ],
     base_image="quay.io/docling-project/docling-serve:jobkit-base-0.0.19",  # base docling-serve image with fixed permissions
 )
 def convert_payload(
     options: dict,
-    source: dict,
     target: dict,
     batch_index: int,
     # source_keys: List[str],
@@ -24,22 +23,12 @@ def convert_payload(
     import json
     import logging
     import os
-    from typing import Optional
 
-    from docling.backend.docling_parse_backend import DoclingParseDocumentBackend
-    from docling.backend.docling_parse_v2_backend import DoclingParseV2DocumentBackend
-    from docling.backend.docling_parse_v4_backend import DoclingParseV4DocumentBackend
-    from docling.backend.pdf_backend import PdfDocumentBackend
-    from docling.backend.pypdfium2_backend import PyPdfiumDocumentBackend
-    from docling.datamodel.pipeline_options import (
-        OcrOptions,
-        PdfBackend,
-        PdfPipelineOptions,
-        TableFormerMode,
+    from docling_jobkit.connectors.s3_helper import ResultsProcessor
+    from docling_jobkit.convert.manager import (
+        DoclingConverterManager,
+        DoclingConverterManagerConfig,
     )
-    from docling.models.factories import get_ocr_factory
-
-    from docling_jobkit.connectors.s3_helper import DoclingConvert
     from docling_jobkit.datamodel.convert import ConvertDocumentsOptions
     from docling_jobkit.datamodel.s3_coords import S3Coordinates
 
@@ -51,67 +40,28 @@ def convert_payload(
     # os.environ["MODULE_PATH"] = str(easyocr_path)
     # os.environ["EASYOCR_MODULE_PATH"] = str(easyocr_path)
 
-    # validate inputs
-    source_s3_coords = S3Coordinates.model_validate(source)
+    # validate coords
     target_s3_coords = S3Coordinates.model_validate(target)
 
     convert_options = ConvertDocumentsOptions.model_validate(options)
 
-    backend: Optional[type[PdfDocumentBackend]] = None
-    if convert_options.pdf_backend:
-        if convert_options.pdf_backend == PdfBackend.DLPARSE_V1:
-            backend = DoclingParseDocumentBackend
-        elif convert_options.pdf_backend == PdfBackend.DLPARSE_V2:
-            backend = DoclingParseV2DocumentBackend
-        elif convert_options.pdf_backend == PdfBackend.DLPARSE_V4:
-            backend = DoclingParseV4DocumentBackend
-        elif convert_options.pdf_backend == PdfBackend.PYPDFIUM2:
-            backend = PyPdfiumDocumentBackend
-        else:
-            raise RuntimeError(
-                f"Unexpected PDF backend type {convert_options.pdf_backend}"
-            )
-
-    pipeline_options = PdfPipelineOptions()
-    pipeline_options.do_ocr = convert_options.do_ocr
-    ocr_factory = get_ocr_factory()
-
-    pipeline_options.ocr_options = cast(
-        OcrOptions, ocr_factory.create_options(kind=convert_options.ocr_engine)
-    )
-
-    pipeline_options.do_table_structure = convert_options.do_table_structure
-    pipeline_options.table_structure_options.mode = TableFormerMode(
-        convert_options.table_mode
-    )
-    pipeline_options.generate_page_images = convert_options.include_images
-    pipeline_options.do_code_enrichment = convert_options.do_code_enrichment
-    pipeline_options.do_formula_enrichment = convert_options.do_formula_enrichment
-    pipeline_options.do_picture_classification = (
-        convert_options.do_picture_classification
-    )
-    pipeline_options.do_picture_description = convert_options.do_picture_description
-    pipeline_options.generate_picture_images = convert_options.include_images
-
-    # pipeline_options.accelerator_options = AcceleratorOptions(
-    #     num_threads=2, device=AcceleratorDevice.CUDA
-    # )
-
-    converter = DoclingConvert(
-        source_s3_coords=source_s3_coords,
-        target_s3_coords=target_s3_coords,
-        pipeline_options=pipeline_options,
-        allowed_formats=[str(v) for v in convert_options.from_formats],
-        to_formats=[str(v) for v in convert_options.to_formats],
-        backend=backend,
-    )
+    config = DoclingConverterManagerConfig()
+    converter = DoclingConverterManager(config)
 
     with open(dataset.path) as f:
         batches = json.load(f)
     source_keys = batches[batch_index]
 
     results = []
-    for item in converter.convert_documents(source_keys):
+    result_processor = ResultsProcessor(
+        target_s3_coords=target_s3_coords,
+        to_formats=[v.value for v in convert_options.to_formats],
+        generate_page_images=convert_options.include_images,
+        generate_picture_images=convert_options.include_images,
+    )
+    for item in result_processor.process_documents(
+        converter.convert_documents(source_keys, options=convert_options)
+    ):
         results.append(item)
         logging.info("Convertion result: {}".format(item))
 
@@ -122,7 +72,7 @@ def convert_payload(
     packages_to_install=[
         "pydantic",
         "boto3~=1.35.36",
-        "git+https://github.com/docling-project/docling-jobkit@2c27c71b75da98f04fccc7abc7ddc3a9a3afb0cd",
+        "git+https://github.com/docling-project/docling-jobkit@main",
     ],
     base_image="python:3.11",
 )
@@ -138,6 +88,7 @@ def compute_batches(
     from docling_jobkit.connectors.s3_helper import (
         check_target_has_source_converted,
         generate_batch_keys,
+        generate_presign_url,
         get_s3_connection,
         get_source_files,
     )
@@ -154,8 +105,12 @@ def compute_batches(
     filtered_source_keys = check_target_has_source_converted(
         s3_target_coords, source_objects_list, s3_coords_source.key_prefix
     )
+    presign_filtered_source_keys = [
+        generate_presign_url(s3_source_client, key, s3_coords_source.bucket)
+        for key in filtered_source_keys
+    ]
     batch_keys = generate_batch_keys(
-        filtered_source_keys,
+        presign_filtered_source_keys,
         batch_size=batch_size,
     )
 
@@ -232,7 +187,6 @@ def inputs_s3in_s3out(
     with dsl.ParallelFor(batches.outputs["batch_indices"], parallelism=20) as subbatch:
         converter = convert_payload(
             options=convertion_options,
-            source=source,
             target=target,
             dataset=batches.outputs["dataset"],
             batch_index=subbatch,